requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-environ==0.9.0
django-filter==2.4.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
